Fixed inspect rotation and patched clipping
Inspect rotation now works relative to the camera instead of the object's actual rotation. Also added a patch thing for objects that manage to clip through the floor; they will be returned to their spawn position after falling for a bit (since the key likes to clip thru floor). Also removed the capsule colliders on the ceiling lights since they were unnecessary and could cause softlocks if player managed to get up on the tables.