Crosshair + prefabs + camera/item position changes
Added a crosshair, texture to change. It is a prefab so it can be imported into any scene. Also changed the player to a prefab and moved its starting position in the tutorial room. Key's position in tutorial room is also shifted.

Player camera position has been changed slightly and so has where objects are held.

Also fixed an autosubmit bug I forgot about from the last commit.